backend/requirements.txt
langchain-text-splitters 
langchain-community 
langgraph
langchain-deepseek
langchain-openai
python-dotenv
langchain-chroma